+++ packages/zowe-explorer/__mocks__/mockCreators/shared.ts
@@ -18,6 +18,9 @@ import { FilterDescriptor } from "../../src/utils/ProfilesUtils";
 import * as zosmf from "@zowe/zosmf-for-zowe-sdk";
 import { SettingsConfig } from "../../src/utils/SettingsConfig";
 import * as globals from "../../src/globals";
+import { Profiles } from "../../src/Profiles";
+
+const MOCK_PROFILES = [];
 
 export function createPersistentConfig() {
     return {
@@ -310,6 +313,7 @@ export function createTextDocument(name: string, sessionNode?: ZoweDatasetNode |
 
 export function createInstanceOfProfile(profile: imperative.IProfileLoaded) {
     return {
+        addToConfigArray: Profiles.prototype.addToConfigArray,
         allProfiles: [{ name: "sestest" }, { name: "profile1" }, { name: "profile2" }],
         defaultProfile: { name: "sestest" },
         getDefaultProfile: jest.fn(),
@@ -321,6 +325,8 @@ export function createInstanceOfProfile(profile: imperative.IProfileLoaded) {
             return { status: "active", name: "sestest" };
         }),
         profilesForValidation: [{ status: "active", name: "sestest" }],
+        profileTypeConfigurations: MOCK_PROFILES,
+        getConfigArray: () => MOCK_PROFILES,
         validateProfiles: jest.fn(),
         getBaseProfile: jest.fn(),
         enableValidationContext: jest.fn(),