+++ packages/zowe-explorer/__tests__/__unit__/dataset/DatasetTree.unit.test.ts
@@ -47,6 +47,7 @@ import { ZoweLocalStorage } from "../../../src/utils/ZoweLocalStorage";
 import { TreeProviders } from "../../../src/shared/TreeProviders";
 import { join } from "path";
 import * as sharedUtils from "../../../src/shared/utils";
+import { mocked } from "../../../__mocks__/mockUtils";
 
 jest.mock("fs");
 jest.mock("util");
@@ -169,9 +170,6 @@ function createGlobalMocks() {
 
     return globalMocks;
 }
-
-// Idea is borrowed from: https://github.com/kulshekhar/ts-jest/blob/eadf408f90ca2007b85f5a61c6c0b74b4e431943/src/utils/testing.ts
-const mocked = <T extends (...args: any[]) => any>(fn: T): jest.Mock<ReturnType<T>> => fn as any;
 
 describe("Dataset Tree Unit Tests - Initialisation", () => {
     function createBlockMocks() {