+++ packages/zowe-explorer/__tests__/__unit__/job/actions.unit.test.ts
@@ -48,6 +48,7 @@ import { ZoweLocalStorage } from "../../../src/utils/ZoweLocalStorage";
 import { LocalFileManagement } from "../../../src/utils/LocalFileManagement";
 import { ProfileManagement } from "../../../src/utils/ProfileManagement";
 import { TreeProviders } from "../../../src/shared/TreeProviders";
+import { mocked } from "../../../__mocks__/mockUtils";
 
 jest.mock("../../../src/utils/LoggerUtils");
 
@@ -154,9 +155,6 @@ function createGlobalMocks() {
 
     return newMocks;
 }
-
-// Idea is borrowed from: https://github.com/kulshekhar/ts-jest/blob/master/src/util/testing.ts
-const mocked = <T extends (...args: any[]) => any>(fn: T): jest.Mock<ReturnType<T>> => fn as any;
 
 afterEach(() => {
     jest.clearAllMocks();